Fix: consertando falhas na apresentação das listas além de otimizar e melhorar a prevenção de erros usando stream

+++ src/main/java/br/com/projetoa3/modelo/records/Notes.java
@@ -2,7 +2,7 @@ package br.com.projetoa3.modelo.records;
 
 public record Notes(int notaA1, int notaA2, int notaA3) {
     public double getMedia() {
-        return (notaA1 + notaA2 + notaA3) / 3;
+        return (double) (notaA1 + notaA2 + notaA3) / 3;
     }
 
     public String getStatus() {
